apps/entity-conformity-endpoints/requirements.txt
fastapi
uvicorn
asyncio-nats-client
pymongo